reorganized prompt pool code
