+++ src/renate/benchmark/models/__init__.py
@@ -9,6 +9,7 @@ from renate.benchmark.models.resnet import (
     ResNet50,
     ResNet50CIFAR,
 )
+from renate.benchmark.models.l2p import PromptedVisionTransformer
 from renate.benchmark.models.vision_transformer import (
     VisionTransformerB16,
     VisionTransformerB32,
@@ -26,6 +27,7 @@ __all__ = [
     "ResNet34CIFAR",
     "ResNet50",
     "ResNet50CIFAR",
+    "PromptedVisionTransformer",
     "VisionTransformerB16",
     "VisionTransformerB32",
     "VisionTransformerCIFAR",
